+++ include/hsmcpp/HsmEventDispatcherArduino.hpp
@@ -25,27 +25,21 @@ private:
 
 public:
     /**
-     * @copydoc HsmEventDispatcherBase::HsmEventDispatcherBase()
+     * @brief Create dispatcher instance.
+     * @param eventsCacheSize size of the queue preallocated for delayed events
+     * @return New dispatcher instance.
+     *
+     * @threadsafe{Instance can be safely created and destroyed from any thread.}
      */
-    explicit HsmEventDispatcherArduino(const size_t eventsCacheSize = DISPATCHER_DEFAULT_EVENTS_CACHESIZE);
-
-    /**
-     * Destructor.
-     */
-    virtual ~HsmEventDispatcherArduino();
+    // cppcheck-suppress misra-c2012-17.8 ; false positive. setting default parameter value is not parameter modification
+    static std::shared_ptr<HsmEventDispatcherArduino> create(
+        const size_t eventsCacheSize = DISPATCHER_DEFAULT_EVENTS_CACHESIZE);
 
     /**
      * @brief See IHsmEventDispatcher::start()
      * @concurrencysafe{ }
      */
     bool start() override;
-
-    /**
-     * @brief Stop dispatching events.
-     * @details Future calls to dispatchEvents() will have no effect.
-     * @concurrencysafe{ }
-     */
-    void stop();
 
     /**
      * @brief See IHsmEventDispatcher::emitEvent()
@@ -63,15 +57,37 @@ public:
     void dispatchEvents();
 
 protected:
+    /**
+     * @copydoc HsmEventDispatcherBase::HsmEventDispatcherBase()
+     */
+    explicit HsmEventDispatcherArduino(const size_t eventsCacheSize);
+
+    /**
+     * Destructor.
+     */
+    virtual ~HsmEventDispatcherArduino();
+
+    /**
+     * @copydoc HsmEventDispatcherBase::deleteSafe()
+     */
+    bool deleteSafe() override;
+
     void notifyDispatcherAboutEvent() override;
 
+    /**
+     * @brief See HsmEventDispatcherBase::startTimerImpl()
+     * @interruptsafe{ }
+     */
     void startTimerImpl(const TimerID_t timerID, const unsigned int intervalMs, const bool isSingleShot) override;
+    /**
+     * @brief See HsmEventDispatcherBase::stopTimerImpl()
+     * @interruptsafe{ }
+     */
     void stopTimerImpl(const TimerID_t timerID) override;
 
     void handleTimers();
 
 private:
-    bool mStopDispatcher = false;
     std::map<TimerID_t, RunningTimerInfo> mRunningTimers;
 };